requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
flask-Bcrypt
pyjwt